plugins/common_utils/requirements.txt
kafka_python
PyYAML